pyproject.toml
[project]
name = "eval-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
